Move beginEditPersistentToken(_:) to MasterPresenter
