+++ Authenticator/Source/MasterPresenter.swift
@@ -24,9 +24,11 @@
 //
 
 import UIKit
+import OneTimePassword
 
 protocol MasterPresenter {
     func beginAddToken()
+    func beginEditPersistentToken(persistentToken: PersistentToken)
 
     func presentViewController(viewController: UIViewController)
     func dismissViewController()
